Add ThumbnailThread to FavouritesModel
I have previously removed the ability to add thumbnails to favourites as
the behaviour is somewhat ambiguous. But I think it is worth doing it,
because user can assign custom, separate thumbnails to saved items.

`process_index()` now takes an additional `exists` argument to check
if the queried file exists. This to remove stale items from the list
of favourites via the thread-workers.