Add thread worker to `AssetsWidget`
I was experiencing performance issues when querying a large-ish number
of assets on the main thread which made me want to take advantage of
parallel processing on Assets as well.

Everything was in-place but some further abstraction was needed:
Added a new `ThreadedBaseWidget` subclass with all the functions needed
to push indexes to the thread workers. The other base classes were
abstracted more as well and now don't explicitly declare their models.