Optimize the delegate.py module
Paint events of items were slow because the delegate were badly
structured. More crucially, I was not saving important information
about where certain drawn parts are that might later need to be
made clickable.

I restructured the old classes, and ironed out a few minor clunks I
found. The appearance is still cluttered but it is more of a design
problem at this point than a technical issue.