Remove `index_update_timer`
Now that the Threaded list widget has its own class it is much
easier to control the signal connections. It, I hope, also means
that there is no longer a need for a brute-force approach to
load index data using a constantly running timer.

I replaced it with normal signal connection responding to
user events.