Use dataSorted instead of modelReset
I didn't take it into account the dataSorted signal.
The main problem was that it was not consistently
closing the __initdata__ sequence: by adding sort_data to
the @initdata decorator I can make sure the set data is
always sorted.

This also means the last signal of a model-load is not
modelReset but dataSorted. The threads and the proxy model methods
need to take this into account.