Improve layout and event handling
The mouse event handling was inconsistent and unreliable. This should
work better now. I also an override cursor to show when the description
rectangle is editable.

This is a major commit with many improvements that I didn't break into
separate commits - they overall should improve the look and feel of the
application.