Implement lock file for multi-sessions
The multisession states are indicated poorly at the moment and often
session mutually override each-other leading to confusing situations.

I'll try to mitigate this by better indicating in the UI the current
SOLO state. Further, if a session has already been initiated, we should
start GWBrowser SOLO turned on: it will be up to the user to set the
state explicitly to track the master instance.

Documentation will have to reflect this clearly too, and I have to
make sure the UI is indicating the state clearly and the possible user
actions as well.