[WIP] Keep implementing the Bookmark widget
There's a new AddTemplateWidget (for adding jobs and assets) I'm working
on. This will depreciate the current mess I made with the server
settings, and template management.

The key is to keep it all under one window to add/remove templates,
and to move the template files to where all the other config files live.
I'll press on soon to get this all sorted out.