Merge Mode, Active and LocalSettings
The Active class was not very meaningful on its own and makes sense
to refer to all of it's functions from withing LocalSettings.
Same if true for the Mode module.

This of course changes how imports are made. I'll eventually import
`gwbrowser.settings` as `settings` but only after AssetSettings
have been removed.