Remove custom __initdata__
One major benefit I did not foresee is that having abstracted some core
functions, I can now do away with the overrides in widget and the model,
which is plendid.

The key to this is that __initdata__ now calls an iterator function that
provides the DirEntries needed to populate our models. Overriding this
to yield saved bookmark entries means we can repopulate our `FilesModel`
with only bookmark data.

I can see this as a way to go forward, down the line, when thinking
about getting not just favourites but, `approved`, `to-be-reviewed`, ..
items.