[WIP] Address data integrity
Whilst indexes won't get invalidated, the underlying model data does get
modified by eg. data sorting. This can lead to an application crash when
a worker tries to access the underlying data that has already been
deleted.

I'm thinking about adding an identifier to the data dict to verify
integrity. And use that in the modelEg. a custom uuid property?