[0.3.0][WIP] Application overhaul
This is a work in progress stage of rewriting how threads are
implemented. It has not yet been finalised - references to methods that
changed or deleted still need to be updated. Especially AssetSettings
and image operations have much code-breaking changes.

There's also lots of other changes, too much to list here, regarding
delegates and the base widgets/models.