[0.3.0][WIP] Work towards a functioning release candidate
There are still big outstanding issues to address but this is the first
take where all prior functions are back and working again.

Much of the codebase has been optimised - including taking a look at
how thumbnails are generated. The new method is about 50% speedier
and utilises the OpenImageIO.ImageCache more effectively. It also
handles transparencies and instead of a solid black background renders
images against a checker pattern.

BOOKMARK_DB TODOs
----------------

Slack is not integrated with bookmarks. The API tokens and the prefix
names should be associated with jobs and should not be global.
There's already a placeholder table for job attributes but this is not
yet implemented.

Modules docs TOOD
-----------------
This will need another pass.

Test Slack and Maya modules
---------------------------
Didn't test the Slack and Maya modules in a long time.

MacOSX
-------
With the new APIs in place I should go ahead and test against MACOSX.