[WIP][0.3.0] Improve codebase
I have been working hard to get the app in shape and to ready it for
further testing. Currently, most modules are autonomous and more or less
function as expected, which is exciting but... bugs galore. The lack of
unit-testing is slowing bug-tracking but I'm not quiet sure how best is
to approach testing. Some library functions such as the core ImageCache
functions and core methods in the common module can probably have their
own unittests.

I have started backporting changes and checking breakage against QT5.6.
I have also started working on the standalone Windows deployment that
have further exposed redundancies. Next up is to get the MacOS release
in shape.

I have ended up creating a new `Bookmark Properties` section where
information about the bookmark will be stored. The data is stored in the
bookmark database and currenty the size, framerate and Slack properties
are implemented. I have also exposed the asset identifier settings here
as this should be a project dependent setting.

I feel like the new setup is suitably flexible for small scale
applications but I have to keep testing.

I have replaced the old ThumbnailViewer with a new viewer that loads the
image data using OpenImageIO and implements a QGrahicsScene for zooming
and panning. This lets us get away with generating smaller thumbnails
and in theory should speed things up.