Improve Slack integration
Admittedly, the implementation is rudimentary and I wasn't taking the
care to read the Slack API docs carefully to understand the role of
Scopes. Also, the API itself keeps shifting so perhaps it was updated
since I first implemented the module.

I was trying to better indicate, via the UI and error messages if
a required Slack Scope is missing.