Address scatter-gun approach to setting stylesheet
I spotted instances of re-setting the stylesheet when it was not
necessary to do so. I had a closer and look and spotted mistakes in
the stylesheet when using DPI scaling. The problem is that since
I'm using pixels as base-values float numbers are no longer values.
Lot's of styling issues go away when changing float to int.