Rewrite images and log modules
IMPORTANT:

There are many changes in this commit that break compatibility with
previous versions. There're core API changes that modify how data is
read and set internally and using the SQLite databases.

There's a more efficient, cache-backed hash method, common.get_hash(),
that the database and the new image cache module uses. There's also a
more careful approach on how pixmaps and qimages are created and
utilised. The image workflow now stems from and OpenImageIO ImageBuf
instance and is based around QImages and is backed by image caches along
the way. QPixmaps are only ever cached and returned in the main GUI
thread (as it is supposed to be done).

I done entirely away with image data inside models, and the flow of
image data relies more heavily on the `images` module independent of
where the data is called. In practice, this make loading thumbnails much
more straightforward because we don't have to worry about keeping the
models updated. It does probably however puts a little burden on the
delegate, because it is here that we try to fingure out which thumbnails
to use when no cache has been saved.

There's a ton of other tweaks too, notably I added a bespoke Log module,
mainly moved code over from common for clarity. The log has now been
added to the preferences widget.