Work on sequence/file path inconsistencies
There are a number of fixes in this commit related to how we're using
proxy paths. The interface for setting item flags was incomplete, or at
least inconsistent. This resulted in the flags not being set correctly
between sequence and file items. I made it clearer also that setting
flags on sequence items will override the flags on file items.

Renamed the baselistwidget module to `baselist` and related to the
proxy_path calvary I fixed how the current selection is saved and
restored.

Also related to the proxy_path usage item thumbnails will now
addittionally will fall back to sequence thumbnails when individual
thumbnails are not available but the sequence has a thumbnail set.

In line with the above I adjusted how the thread worker loads item
flags. There're a number of other tweaks related to errors spotted
whilst working on the documentation.