Store active paths only in the settings module
It is an entirely pointless excersie storing active path selections
in models as this value is always global to the current session of Bookmarks.
Probably best to always refer to the `settings.ACTIVE` dictionary where
the active path selections are stored. The dictionary is syncronised with the
local settings ini file where the selections are stored long term.