Rework internals to allow saving excat properties
This has been bugging me for a while, when a filter or a flag was
set for an item in folder/x this filder would still be applied for
all folders called 'x'. Instead of this now filters will be specific
to the actual folder allowing content filters to be more customized.

This meant that the keys used to save the filters and flags now are
provided by dynamic methods using the active path selections.