WIP Shotgun intergration
Extended the bookmark_db with new shotgun specific columns.
For the intergration to fully operate more work will have to be
done and testing is missing at the moment.

But. The current setup allows script/api users to associate local
jobs/asset with shotgun projects/assets.
The integrations are saved in the bookmark_db but the current values are
probably will be subject to a review once I wrapped my head around
the shotgun API.