Add custom field to default file-name
This is something I should have fixed before - the custom name
segment should always be part of the generated name. When not specified
it defaults to 'main' - just a generic term, otherwise this is the
segment to descripbe subelements within an asset or shot task. Eg.
'foreground', 'background', 'windows', etc.