Add export for OBJ, Arnold ASS and selector for ScreenCapture
I'm replacing some precoded elements because I realised it's possible to
query the Workspace to get the path to an export folder. Hence, instead
of offering custom paths via preferences it is better to just serve what
is defined by the current workspace.

I also added a panel selector for the capture process - I was previously
relying on getting the active panel but this proved to be totally
unreliable because the panel would lose focus as soon as the user would
select the capture option in bookmarks. Instead,  we'll let to user pick
a model panel currently visible prior to starting the capture process.