Change package imports
The package naming is somewhat inconsistent. To address this,
removed  `widget` from the module names and prepended the list-related
widgets with `list*`. The base list module is now called `lists.py`.

The aim is to remove absolute imports and use relative imports across
the project. This will allow us to fork and load multiple versions
of bookmarks to the same environment without changing the source
code for each fork drastically.