Leave saved case as it is
It is entirely my fault for not giving ot more thought and as a result
letter-case handling in the interiours are messy and inconsistent.

Initially, I was thinking it is fine if I convert all my saved strings
to lower-case as working on windows we're not bound by case dependency.
However, I started noticing strange behaviour in Maya where actually, it
might be better if we didn't mess with lower and upper-cases. In any
case, the right path would have been to save everything without
modification and convert the letter cases on a need-to-basis.