Work in progress edits
I have fallen a bit behind on the git side of things and I haven't been
commiting to the master branch as much as I should have.

There's a series of parallel changes and refractoring contained in this
commit. The main changes concert declaring defaults in a more trasparent
manner.

The bookmark_db table definitions are now declared in a
bookmark_db.json, and the default paths module is making its way into
the bookmark config. The data and the editors are being worked on in the
`asset_config` submodule.