Work in progress actions.py module
User actions are not defined cohesively in one central module but
instead are scattered around the `common`, `contextmenu`, `lists.base`
and other submodules. To implement shortcuts, and more manageable
unittests, I'm moving as much of the action definitions to this new
module. Still a work-in-progress but having done some work on this
I can see lots of room for improvement in the old code-base.
I'm kind of excited to implement this!